Optimize the project structure and add namespaces
