+++ Makefile
@@ -1,13 +1,23 @@
-GPPPARAMS = -m32 -fno-use-cxa-atexit -nostdlib -fno-builtin -fno-rtti -fno-exceptions -fno-leading-underscore -Wno-write-string
+GPPPARAMS = -m32 -Iinclude -fno-use-cxa-atexit -nostdlib -fno-builtin -fno-rtti -fno-exceptions -fno-leading-underscore -Wno-write-string
 ASPARAMS = --32
 LDPARAMS = -melf_i386
 
-objects = loader.o gdt.o driver.o port.o interruptstubs.o interrupts.o keyboard.o mouse.o kernel.o
+objects = 	obj/loader.o \
+			obj/gdt.o \
+			obj/drivers/driver.o \
+			obj/hardwarecommunication/port.o \
+			obj/hardwarecommunication/interruptstubs.o \
+			obj/hardwarecommunication/interrupts.o \
+			obj/drivers/keyboard.o \
+			obj/drivers/mouse.o \
+			obj/kernel.o
 
-%.o : %.cpp
+obj/%.o : src/%.cpp
+	mkdir -p $(@D)
 	g++ $(GPPPARAMS) -o $@ -c $<
 
-%.o : %.s
+obj/%.o : src/%.s
+	mkdir -p $(@D)
 	as $(ASPARAMS) -o $@ $<
 
 GTOS.bin : linker.ld $(objects)
@@ -36,4 +46,4 @@ run : GTOS.iso
 	VirtualBoxVM --startvm GTOS &
 
 rm : GTOS.iso
-	rm -f *.o *.bin *.iso
+	rm -rf obj GTOS.bin GTOS.iso
