+++ include/common/types.h
@@ -1,5 +1,7 @@
-#ifndef __TYPES_H
-#define __TYPES_H
+#ifndef __GTOS__COMMON__TYPES_H
+#define __GTOS__COMMON__TYPES_H
+
+
 
     typedef char int8_t;
     typedef unsigned char uint8_t;
