+++ include/drivers/driver.h
@@ -1,0 +1,30 @@
+#ifndef __GTOS__DRIVERS__DRIVER_H
+#define __GTOS__DRIVERS__DRIVER_H
+
+namespace gtos {
+    namespace drivers {
+        class Driver {
+            public:
+                Driver();
+                ~Driver();
+
+                virtual void Activate();
+                virtual int Reset();
+                virtual void Deactivate();
+        };
+
+        class DriverManager {
+            private:
+                Driver* drivers[255];
+                int numDrivers;
+            public:
+                DriverManager();
+                void AddDriver(Driver*);
+                void ActivateAll();
+        };
+
+    }
+}
+
+
+#endif
