+++ src/drivers/driver.cpp
@@ -1,4 +1,6 @@
-#include "driver.h"
+#include <drivers/driver.h>
+
+using namespace gtos::drivers;
 
 Driver::Driver() {
